requirements.txt
numpy==1.21.4
scikit-learn==1.1.1
streamlit==1.18.0
Pillow==9.2.0
